% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{convertEnsambleToSymbol}
\alias{convertEnsambleToSymbol}
\title{convertEnsambleToSymbol}
\usage{
convertEnsambleToSymbol(gene_list)
}
\arguments{
\item{gene_list}{}
}
\value{

}
\description{
convertEnsambleToSymbol
}
