% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{correlationOmics}
\alias{correlationOmics}
\title{Correlation beetween omics data}
\usage{
correlationOmics(dataOmics1, dataOmics2)
}
\arguments{
\item{dataOmics1}{}

\item{dataOmics2}{}
}
\value{

}
\description{
Get correlation of genes between different omics data
}
