% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getDataType}
\alias{getDataType}
\title{Get omics data type}
\usage{
getDataType(CancerType, DataType)
}
\arguments{
\item{CancerType}{}

\item{DataType}{}
}
\value{

}
\description{
Allow to download one type of omics data from the dataset
}
\details{
"Clinical"        "SCNV"            "Mutation"        "Acetylproteome"  "Phosphoproteome" "Proteome"        "RNAseq"
"miRNA"           "Circular"        "Methylation"     "CNV"             "Glycoproteome"   "RPPA"
}
