% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getOmicsList}
\alias{getOmicsList}
\title{Omics datasets list}
\usage{
getOmicsList(CancerType)
}
\arguments{
\item{CancerType}{}
}
\value{

}
\description{
Returns the list of available omics data of cancer
}
