% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Plot correlation}
\usage{
plotCorrelation(correlation_df)
}
\arguments{
\item{correlation_df}{}
}
\value{

}
\description{
Plot correlation of genes between different omics data
}
