% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{plotMutations}
\alias{plotMutations}
\title{Plot mutations}
\usage{
plotMutations(Mutation, top = 30)
}
\arguments{
\item{Mutation}{}

\item{top}{}
}
\value{

}
\description{
Heatmap mutations or oncoplot of annotated mutations
}
