% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{subsetSharedFetures}
\alias{subsetSharedFetures}
\title{Title}
\usage{
subsetSharedFetures(RNAseq, Mutation, gene_name)
}
\arguments{
\item{RNAseq}{}

\item{Mutation}{}

\item{gene_name}{}
}
\value{

}
\description{
Title
}
