% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{subsetSharedMutation}
\alias{subsetSharedMutation}
\title{Get omics data with shared mutation}
\usage{
subsetSharedMutation(Data, Mutation, gene_name)
}
\arguments{
\item{Data}{}

\item{Mutation}{}

\item{gene_name}{}
}
\value{

}
\description{
Subset omics data with shared mutation
}
