% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{topMutatedGenes}
\alias{topMutatedGenes}
\title{Get top mutated genes}
\usage{
topMutatedGenes(Mutation, num_genes = 30)
}
\arguments{
\item{Mutation}{}

\item{num_genes}{}
}
\value{

}
\description{
Return list of genes more mutated
}
