% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getData}
\alias{getData}
\title{Get omics data}
\usage{
getData(CancerType, DataType)
}
\arguments{
\item{CancerType}{}

\item{DataType}{}
}
\value{

}
\description{
Allow to download one or more omics data from the dataset
}
