% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Plot correlation}
\usage{
plotCorrelation(dataOmics1, dataOmics2)
}
\arguments{
\item{dataOmics1}{}

\item{dataOmics2}{}
}
\value{

}
\description{
Plot correlation of genes between different omics data
}
